[state] move power control to separate compilation unit
